requirements.txt
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
keras>=3.0.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
torch==2.1.0
torchvision==0.16.0
opencv-python-headless==4.8.1.78
pillow==10.1.0
numpy==1.24.3